✨ Introduce BundleVersionMatcher and BundleUtilities, consolidate InstallationCoordinator into UpdateInstaller
