+++ Loop/Updater/FileVerifier.swift
@@ -73,62 +73,4 @@ class FileVerifier {
         Log.debug("SHA256 calculation complete - Checksum: \(checksum), File: \(fileURL.lastPathComponent)")
         return checksum
     }
-
-    private func findAppBundle(in directory: URL) throws -> URL {
-        let contents = try FileManager.default.contentsOfDirectory(
-            at: directory,
-            includingPropertiesForKeys: [.isDirectoryKey]
-        )
-
-        for item in contents {
-            if item.pathExtension == "app" {
-                Log.debug("Found app bundle - Path: \(item.path)")
-                return item
-            }
-
-            let resourceValues = try item.resourceValues(forKeys: [.isDirectoryKey])
-            if resourceValues.isDirectory == true {
-                if let found = try? findAppBundle(in: item) {
-                    return found
-                }
-            }
-        }
-
-        throw UpdateError.installationError("Could not find .app bundle in extracted files")
-    }
-
-    private func verifyBundleStructure(_ appBundle: URL) throws {
-        let requiredPaths = ["Contents/Info.plist", "Contents/MacOS"]
-
-        for path in requiredPaths {
-            let fullPath = appBundle.appendingPathComponent(path)
-            guard FileManager.default.fileExists(atPath: fullPath.path) else {
-                throw UpdateError.installationError("Missing required bundle component: \(path)")
-            }
-        }
-
-        Log.info("Bundle structure verification passed")
-    }
-
-    private func verifyCodeSignature(_ appBundle: URL) async throws {
-        Log.debug("Verifying code signature")
-
-        let process = Process()
-        process.executableURL = URL(fileURLWithPath: "/usr/bin/codesign")
-        process.arguments = ["--verify", "--verbose", appBundle.path]
-
-        let pipe = Pipe()
-        process.standardError = pipe
-
-        try process.run()
-        process.waitUntilExit()
-
-        if process.terminationStatus != 0 {
-            let data = pipe.fileHandleForReading.readDataToEndOfFile()
-            let output = String(data: data, encoding: .utf8) ?? "Unknown error"
-            Log.warn("Code signature verification failed: \(output)")
-        } else {
-            Log.info("Code signature verification successful")
-        }
-    }
 }
